Remove redundant lattice fixture definitions in STQBM tests
