+++ test/unit/spacetime/d1q2_block_test.py
@@ -4,93 +4,6 @@ from qlbm.lattice.geometry.shapes.block import Block
 from qlbm.lattice.lattices.spacetime_lattice import SpaceTimeLattice
 
 
-@pytest.fixture
-def simple_1d_block() -> Block:
-    return Block([(5, 11)], [4], "bounceback")
-
-
-@pytest.fixture
-def simple_large_1d_block() -> Block:
-    return Block([(2, 14)], [4], "bounceback")
-
-
-@pytest.fixture
-def dummy_lattice() -> SpaceTimeLattice:
-    return SpaceTimeLattice(
-        0,
-        {
-            "lattice": {
-                "dim": {"x": 256},
-                "velocities": {"x": 2},
-            },
-        },
-    )
-
-
-@pytest.fixture
-def lattice_1d_16_1_obstacle_1_timestep() -> SpaceTimeLattice:
-    return SpaceTimeLattice(
-        1,
-        {
-            "lattice": {
-                "dim": {"x": 16},
-                "velocities": {"x": 2},
-            },
-            "geometry": [
-                {"x": [5, 11], "boundary": "bounceback"},
-            ],
-        },
-    )
-
-
-@pytest.fixture
-def lattice_1d_16_1_obstacle_2_timesteps() -> SpaceTimeLattice:
-    return SpaceTimeLattice(
-        2,
-        {
-            "lattice": {
-                "dim": {"x": 16},
-                "velocities": {"x": 2},
-            },
-            "geometry": [
-                {"x": [5, 11], "boundary": "bounceback"},
-            ],
-        },
-    )
-
-
-@pytest.fixture
-def lattice_1d_16_1_obstacle_5_timesteps() -> SpaceTimeLattice:
-    return SpaceTimeLattice(
-        5,
-        {
-            "lattice": {
-                "dim": {"x": 16},
-                "velocities": {"x": 2},
-            },
-            "geometry": [
-                {"x": [5, 11], "boundary": "bounceback"},
-            ],
-        },
-    )
-
-
-@pytest.fixture
-def lattice_1d_16_1_obstacle_5_timesteps_large_obstacle() -> SpaceTimeLattice:
-    return SpaceTimeLattice(
-        5,
-        {
-            "lattice": {
-                "dim": {"x": 16},
-                "velocities": {"x": 2},
-            },
-            "geometry": [
-                {"x": [2, 14], "boundary": "bounceback"},
-            ],
-        },
-    )
-
-
 def test_get_spacetime_reflection_data_constructor_1_timestep(
     simple_1d_block, lattice_1d_16_1_obstacle_1_timestep
 ):
