+++ test/unit/spacetime/d2q4_block_volumetric_test.py
@@ -1,80 +1,14 @@
 from typing import List
+
 import pytest
 
-from qlbm.lattice.geometry.shapes.block import Block
 from qlbm.lattice.geometry.encodings.spacetime import SpaceTimeVolumetricReflectionData
-from qlbm.lattice.lattices.spacetime_lattice import SpaceTimeLattice
-from qlbm.tools.utils import flatten, get_qubits_to_invert
+from qlbm.lattice.geometry.shapes.block import Block
 
 
 @pytest.fixture
 def simple_2d_block():
     return Block([(5, 6), (2, 10)], [4, 4], "bounceback")
-
-
-@pytest.fixture
-def dummy_lattice() -> SpaceTimeLattice:
-    return SpaceTimeLattice(
-        0,
-        {
-            "lattice": {
-                "dim": {"x": 32, "y": 32},
-                "velocities": {"x": 2, "y": 2},
-            },
-        },
-    )
-
-
-@pytest.fixture
-def lattice_2d_16x16_1_obstacle_1_timestep() -> SpaceTimeLattice:
-    return SpaceTimeLattice(
-        1,
-        {
-            "lattice": {
-                "dim": {"x": 16, "y": 16},
-                "velocities": {"x": 2, "y": 2},
-            },
-        },
-    )
-
-
-@pytest.fixture
-def lattice_2d_16x16_1_obstacle_2_timesteps() -> SpaceTimeLattice:
-    return SpaceTimeLattice(
-        2,
-        {
-            "lattice": {
-                "dim": {"x": 16, "y": 16},
-                "velocities": {"x": 2, "y": 2},
-            },
-        },
-    )
-
-
-@pytest.fixture
-def lattice_2d_16x16_1_obstacle_5_timesteps() -> SpaceTimeLattice:
-    return SpaceTimeLattice(
-        5,
-        {
-            "lattice": {
-                "dim": {"x": 16, "y": 16},
-                "velocities": {"x": 2, "y": 2},
-            },
-        },
-    )
-
-
-@pytest.fixture
-def lattice_1d_1x6_1_obstacle_5_timesteps_large_obstacle() -> SpaceTimeLattice:
-    return SpaceTimeLattice(
-        5,
-        {
-            "lattice": {
-                "dim": {"x": 16, "y": 16},
-                "velocities": {"x": 2, "y": 2},
-            },
-        },
-    )
 
 
 def test_volumetric_reflection_data_1_step(
