+++ test/unit/spacetime/d2q4_circle_test.py
@@ -1,0 +1,9 @@
+import pytest
+
+from qlbm.lattice.geometry.shapes.block import Block
+from qlbm.lattice.lattices.spacetime_lattice import SpaceTimeLattice
+from qlbm.tools.utils import flatten, get_qubits_to_invert
+
+
+def test_circle_easy():
+    pass
